+++ tests/testthat.R
@@ -1,0 +1,5 @@
+if (requireNamespace("testthat", quietly = TRUE)) {
+  library(testthat)
+  library(UlavalSSD)
+  test_check("UlavalSSD")
+}